requirements.txt
pytesseract
pdf2image
Pillow
google-generativeai
pandas
opencv-python
numpy
PyPDF2